refactor(alerts): split workspace and improve location resolution
